converted TeacherDetailPage
